feat: F3 — extract hover, completion, signature_help into features layer
- features/completion.rs: compute_completions(&impl CompletionIndex)
- features/hover.rs: compute_hover<W: WorkspaceRead>(workspace, ctx, uri, pos)
- features/signature_help.rs: compute_signature_help(&impl DocumentAccess + SignatureIndex + CallInfoAccess)
- features/traits.rs: add CallInfoAccess trait (CST-only, separate from SignatureIndex)
- features/traits_impl.rs: impl CallInfoAccess for Indexer
- backend/handlers.rs: hover_impl + signature_help_impl → thin adapters (~5 lines each)
- backend/actions.rs: completion_impl → thin adapter (~5 lines)
- backend/mod.rs: remove Backend::resolve_with_receiver_fallback (moved to features/hover)
- backend/format.rs, actions.rs: pub(super) → pub(crate) for cross-module access
