+++ src/backend/handlers.rs
@@ -1,13 +1,6 @@
-use super::actions::is_non_call_keyword;
 use super::cursor::CursorContext;
-use super::format::{format_contextual_hover, format_symbol_hover};
 use super::Backend;
-use crate::indexer::resolution::{
-    enrich_at_location, resolve_symbol_info, ResolveOptions, SubstitutionContext, WorkspaceRead,
-};
-use crate::indexer::{apply_type_subst, cst_call_info, find_fun_signature_with_receiver, CallInfo};
 use crate::inlay_hints::compute_inlay_hints;
-use crate::StrExt;
 use tower_lsp::jsonrpc::Result;
 use tower_lsp::lsp_types::*;
 
@@ -25,117 +18,9 @@ impl Backend {
             return Ok(None);
         };
 
-        if let Some(hover) = self.contextual_lambda_hover(workspace, &ctx, uri, position) {
-            return Ok(Some(hover));
-        }
-        if ctx.qualifier.is_none() && ctx.lambda_decl.is_some() {
-            return Ok(None);
-        }
-        if let Some(hover) = self.contextual_receiver_hover(workspace, &ctx, uri, position) {
-            return Ok(Some(hover));
-        }
-
-        Ok(self.regular_symbol_hover(workspace, &ctx, uri, position))
-    }
-
-    fn contextual_lambda_hover<W: WorkspaceRead>(
-        &self,
-        workspace: &W,
-        ctx: &CursorContext,
-        uri: &Url,
-        position: Position,
-    ) -> Option<Hover> {
-        if ctx.qualifier.is_some() {
-            return None;
-        }
-        let receiver_type = ctx.contextual.as_ref()?;
-        let type_name =
-            self.contextual_hover_type_name(workspace, receiver_type, uri, position.line);
-        let leaf = type_name.rsplit('.').next().unwrap_or(type_name.as_str());
-        let signature = format!("{} {}: {type_name}", hover_binding_keyword(uri), ctx.word);
-        let detail = self
-            .resolve_hover_markdown(workspace, leaf, None, uri, position.line)
-            .or_else(|| crate::stdlib::hover(leaf));
-        Some(make_markdown_hover(format_contextual_hover(
-            &signature,
-            uri.path(),
-            detail.as_deref(),
-        )))
-    }
-
-    fn contextual_hover_type_name<W: WorkspaceRead>(
-        &self,
-        workspace: &W,
-        receiver_type: &crate::resolver::ReceiverType,
-        uri: &Url,
-        line: u32,
-    ) -> String {
-        let subst = crate::indexer::resolution::build_subst_map(workspace, uri.as_str(), line);
-        if subst.is_empty() {
-            return receiver_type.raw.clone();
-        }
-        apply_type_subst(&receiver_type.raw, &subst)
-    }
-
-    fn contextual_receiver_hover<W: WorkspaceRead>(
-        &self,
-        workspace: &W,
-        ctx: &CursorContext,
-        uri: &Url,
-        position: Position,
-    ) -> Option<Hover> {
-        let receiver_type = ctx.contextual.as_ref()?;
-        ctx.qualifier.as_ref()?;
-        let location = self
-            .resolve_with_receiver_fallback(workspace, &ctx.word, receiver_type, uri)
-            .first()?
-            .clone();
-        let info = enrich_at_location(
-            workspace,
-            &location,
-            &ctx.word,
-            hover_substitution_context(uri, position.line),
-            &ResolveOptions::hover(),
-        )?;
-        Some(make_markdown_hover(format_symbol_hover(&info, uri.path())))
-    }
-
-    fn regular_symbol_hover<W: WorkspaceRead>(
-        &self,
-        workspace: &W,
-        ctx: &CursorContext,
-        uri: &Url,
-        position: Position,
-    ) -> Option<Hover> {
-        let markdown = self
-            .resolve_hover_markdown(
-                workspace,
-                &ctx.word,
-                ctx.qualifier.as_deref(),
-                uri,
-                position.line,
-            )
-            .or_else(|| crate::stdlib::hover(&ctx.word))?;
-        Some(make_markdown_hover(markdown))
-    }
-
-    fn resolve_hover_markdown<W: WorkspaceRead>(
-        &self,
-        workspace: &W,
-        word: &str,
-        qualifier: Option<&str>,
-        uri: &Url,
-        line: u32,
-    ) -> Option<String> {
-        resolve_symbol_info(
-            workspace,
-            word,
-            qualifier,
-            uri,
-            hover_substitution_context(uri, line),
-            &ResolveOptions::hover(),
-        )
-        .map(|info| format_symbol_hover(&info, uri.path()))
+        Ok(crate::features::hover::compute_hover(
+            workspace, &ctx, uri, position,
+        ))
     }
 
     pub(super) async fn references_impl(
@@ -263,43 +148,10 @@ impl Backend {
     ) -> Result<Option<SignatureHelp>> {
         let uri = &params.text_document_position_params.text_document.uri;
         let pos = params.text_document_position_params.position;
-
-        // Use live_lines for the current line (updated synchronously on every
-        // keystroke) so signatureHelp fires immediately when `(` is typed,
-        // without waiting for the 120ms debounce that updates `files`.
-        let Some(lines_owned) = self.indexer.mem_lines_for(uri.as_str()) else {
-            return Ok(None);
-        };
-        let lines: &[String] = &lines_owned;
-
-        let line_idx = pos.line as usize;
-        if line_idx >= lines.len() {
-            return Ok(None);
-        }
-        let line_text = &lines[line_idx];
-        // pos.character is UTF-16 units — convert to a byte offset.
-        let col = crate::indexer::live_tree::utf16_col_to_byte(line_text, pos.character as usize);
-        let before = &line_text[..col];
-
-        // Extract CallInfo — CST first, text fallback.
-        let Some(ci) = extract_call_info(pos, &self.indexer, uri, lines, before, line_idx) else {
-            return Ok(None);
-        };
-
-        let params_text = find_fun_signature_with_receiver(
-            &self.indexer,
+        Ok(crate::features::signature_help::compute_signature_help(
             uri,
-            &ci.fn_name,
-            ci.qualifier.as_deref(),
-        );
-        if params_text.is_empty() {
-            return Ok(None);
-        }
-
-        Ok(build_signature_help(
-            &ci.fn_name,
-            &params_text,
-            ci.active_param,
+            pos,
+            self.indexer.as_ref(),
         ))
     }
 
@@ -429,216 +281,6 @@ impl Backend {
     }
 }
 
-// ─── Private helpers for signature_help_impl ─────────────────────────────────
-
-/// Build a `SignatureHelp` response from pre-computed parts.
-fn build_signature_help(
-    fn_name: &str,
-    params_text: &str,
-    active_param: u32,
-) -> Option<SignatureHelp> {
-    let raw = params_text.trim_matches(|c| c == '(' || c == ')');
-    let param_parts: Vec<&str> = raw
-        .split(',')
-        .map(|s| s.trim())
-        .filter(|s| !s.is_empty())
-        .collect();
-    let parameters: Vec<ParameterInformation> = param_parts
-        .iter()
-        .map(|p| ParameterInformation {
-            label: ParameterLabel::Simple(p.to_string()),
-            documentation: None,
-        })
-        .collect();
-    let label = format!("{}({})", fn_name, param_parts.join(", "));
-    let active_param = active_param.min(parameters.len().saturating_sub(1) as u32);
-    Some(SignatureHelp {
-        signatures: vec![SignatureInformation {
-            label,
-            documentation: None,
-            parameters: Some(parameters),
-            active_parameter: Some(active_param),
-        }],
-        active_signature: Some(0),
-        active_parameter: Some(active_param),
-    })
-}
-
-/// Extract [`CallInfo`] for the call under the cursor.
-///
-/// Tries the CST (live tree) first — O(depth), accurate qualifier extraction.
-/// Falls back to a text scan when no live tree is available, when the cursor
-/// is inside a lambda literal, or when the callee shape is not recognised.
-fn extract_call_info(
-    pos: Position,
-    indexer: &crate::indexer::Indexer,
-    uri: &Url,
-    lines: &[String],
-    before: &str,
-    line_idx: usize,
-) -> Option<CallInfo> {
-    if let Some(result) = cst_call_info(pos, indexer, uri) {
-        return Some(result);
-    }
-
-    text_call_info(lines, before, line_idx)
-}
-
-/// Scans a single source line for an unclosed call-site opening.
-/// Returns `(call_name, qualifier)` if an unbalanced `name(` is found,
-/// where net > 0 means more opens than closes on this line.
-fn find_call_open_on_line(line: &str) -> Option<(String, Option<String>)> {
-    for (p, _) in line
-        .char_indices()
-        .filter(|&(_, c)| c == '(')
-        .collect::<Vec<_>>()
-        .into_iter()
-        .rev()
-    {
-        let before_paren = &line[..p];
-        let name = before_paren.last_ident_in();
-        if !name.is_empty() && !is_non_call_keyword(name) {
-            let net: i32 = line[p..]
-                .chars()
-                .map(|c| match c {
-                    '(' => 1,
-                    ')' => -1,
-                    _ => 0,
-                })
-                .sum();
-            if net > 0 {
-                // Qualifier before the dot on the same line.
-                let before_name = &before_paren[..before_paren.len() - name.len()];
-                let qualifier = if before_name.ends_with('.') {
-                    let q = before_name
-                        .strip_suffix('.')
-                        .unwrap_or(before_name)
-                        .last_ident_in();
-                    if q.is_empty() {
-                        None
-                    } else {
-                        Some(q.to_owned())
-                    }
-                } else {
-                    None
-                };
-                return Some((name.to_owned(), qualifier));
-            }
-        }
-    }
-    None
-}
-
-/// Scans up to `MAX_SCAN_BACK_LINES` lines before `line_idx` for an unclosed `fn(` call site.
-/// Returns `(call_name, qualifier, extra_commas)` where `extra_commas` counts commas on the
-/// intermediate lines only (between the opening line and `line_idx`, exclusive). Commas on
-/// `line_idx` itself (in `before`) are already counted by the caller.
-/// Maximum number of lines to scan backward when looking for a multi-line call opener.
-const MAX_SCAN_BACK_LINES: usize = 10;
-
-fn scan_multiline_call_open(
-    lines: &[String],
-    line_idx: usize,
-) -> Option<(String, Option<String>, u32)> {
-    let scan_start = line_idx.saturating_sub(MAX_SCAN_BACK_LINES);
-    for scan_line in (scan_start..line_idx).rev() {
-        let l = &lines[scan_line];
-        if l.contains('{') || l.contains('}') {
-            break;
-        }
-        if let Some((name, qualifier)) = find_call_open_on_line(l) {
-            let mut extra: u32 = 0;
-            if scan_line + 1 < line_idx {
-                for mid in &lines[(scan_line + 1)..line_idx] {
-                    extra += mid.chars().filter(|&c| c == ',').count() as u32;
-                }
-            }
-            return Some((name, qualifier, extra));
-        }
-    }
-    None
-}
-
-/// Given `chars` and position `j` (start of the identifier), extract
-/// the qualifier immediately before a `.` if present.
-fn extract_dot_qualifier(chars: &[char], j: usize) -> Option<String> {
-    if j > 0 && chars[j - 1] == '.' {
-        let mut k = j - 1;
-        while k > 0 && (chars[k - 1].is_alphanumeric() || chars[k - 1] == '_') {
-            k -= 1;
-        }
-        let q: String = chars[k..j - 1].iter().collect();
-        if !q.is_empty() {
-            Some(q)
-        } else {
-            None
-        }
-    } else {
-        None
-    }
-}
-
-/// Text-scan fallback: extract `(fn_name, qualifier, active_param)` by walking
-/// backwards through `before` (and up to 10 previous lines for multiline calls).
-fn text_call_info(lines: &[String], before: &str, line_idx: usize) -> Option<CallInfo> {
-    let mut depth: i32 = 0;
-    let mut active_param: u32 = 0;
-    let mut call_name: Option<String> = None;
-    let mut call_qualifier: Option<String> = None;
-
-    let chars: Vec<char> = before.chars().collect();
-    let mut i = chars.len();
-    while i > 0 {
-        i -= 1;
-        match chars[i] {
-            ')' | ']' => {
-                depth += 1;
-            }
-            '{' | '}' => {
-                break;
-            }
-            '(' => {
-                if depth == 0 {
-                    let mut j = i;
-                    while j > 0 && (chars[j - 1].is_alphanumeric() || chars[j - 1] == '_') {
-                        j -= 1;
-                    }
-                    let candidate: String = chars[j..i].iter().collect();
-                    if !candidate.is_empty() && !is_non_call_keyword(&candidate) {
-                        call_name = Some(candidate);
-                        call_qualifier = extract_dot_qualifier(&chars, j);
-                    }
-                    break;
-                }
-                depth -= 1;
-            }
-            ',' if depth == 0 => {
-                active_param += 1;
-            }
-            _ => {}
-        }
-    }
-
-    // Multiline: scan up to 10 lines back when the call opens on a previous line.
-    let in_block_body = before.contains('{')
-        || before.contains('}')
-        || lines[line_idx].trim_start().starts_with('}');
-    if call_name.is_none() && line_idx > 0 && !in_block_body {
-        if let Some((name, qual, extra)) = scan_multiline_call_open(lines, line_idx) {
-            call_name = Some(name);
-            call_qualifier = qual;
-            active_param += extra;
-        }
-    }
-
-    let fn_name = call_name.filter(|n| !n.is_empty())?;
-    Some(CallInfo {
-        fn_name,
-        qualifier: call_qualifier,
-        active_param,
-    })
-}
-
 #[derive(Clone)]
 struct WorkspaceSymbolQuery {
     raw: String,
@@ -729,27 +371,6 @@ fn rg_workspace_symbol(name: String, location: Location) -> SymbolInformation {
         deprecated: None,
         location,
         container_name: Some("rg fallback".to_string()),
-    }
-}
-
-fn hover_binding_keyword(uri: &Url) -> &'static str {
-    crate::Language::from_path(uri.path()).val_keyword()
-}
-
-fn hover_substitution_context(uri: &Url, line: u32) -> SubstitutionContext<'_> {
-    SubstitutionContext::CrossFile {
-        calling_uri: uri.as_str(),
-        cursor_line: Some(line),
-    }
-}
-
-fn make_markdown_hover(markdown: String) -> Hover {
-    Hover {
-        contents: HoverContents::Markup(MarkupContent {
-            kind: MarkupKind::Markdown,
-            value: markdown,
-        }),
-        range: None,
     }
 }
